Updated Protection Board code
Added basic coulomb counter and full charge pack voltage recalibration for the counter. When the pack has his the full pack voltage value set in the parameter then the Ah of the pack will reset to entered Ah of pack.  Need to optimize later down the road but this is good enough for most people.